requirements.txt
pytest==8.3.5
pytest-html==4.1.1
pytest-xdist==3.6.1
selenium==4.29.0
python-dotenv==1.0.1